How-to section in progress
How-to-section development in progress.

What is done:
Basic layout finished. Sample photos are implemented, texts are written, and heading for how-to section is in progress

What needs to be done:
Completion of how-to section overall;  subheading, images, background and font size needs to be correctly implemented. Also needs overall adjustment.